python-stack/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pyyaml==6.0.2
argon2-cffi==23.1.0
python-dotenv==1.0.1
